+++ Makefile
@@ -66,6 +66,7 @@ install: libdragon.a libdragonsys.a
 	install -m 0644 include/debug.h $(INSTALLDIR)/mips64-elf/include/debug.h
 	install -m 0644 include/usb.h $(INSTALLDIR)/mips64-elf/include/usb.h
 	install -m 0644 include/console.h $(INSTALLDIR)/mips64-elf/include/console.h
+	install -m 0644 include/joybus.h $(INSTALLDIR)/mips64-elf/include/joybus.h
 	install -m 0644 include/mempak.h $(INSTALLDIR)/mips64-elf/include/mempak.h
 	install -m 0644 include/controller.h $(INSTALLDIR)/mips64-elf/include/controller.h
 	install -m 0644 include/rtc.h $(INSTALLDIR)/mips64-elf/include/rtc.h
