+++ include/libdragon.h
@@ -47,6 +47,7 @@
 #include "audio.h"
 #include "console.h"
 #include "debug.h"
+#include "joybus.h"
 #include "controller.h"
 #include "rtc.h"
 #include "mempak.h"
